refactor(crypto): extract shared __rt_digest_to_string runtime helper
