+++ src/codegen/runtime/strings/mod.rs
@@ -67,6 +67,7 @@ mod md5;
 mod sha1;
 mod crc32;
 mod hash;
+mod digest_to_string;
 mod sscanf;
 mod rtrim_mask;
 mod ltrim_mask;
@@ -189,6 +190,8 @@ pub use crc32::emit_crc32;
 /// Emit SHA1 hash helper.
 pub use hash::emit_hash;
 /// Emit generic hash helper.
+pub use digest_to_string::emit_digest_to_string;
+/// Emit the shared raw-digest-to-PHP-string formatter used by hash/md5/sha1.
 pub use sscanf::emit_sscanf;
 /// Emit string scanf parsing helper.
 pub use rtrim_mask::emit_rtrim_mask;
